fix: change from .js to .jsx
