--- src/components/Garage/index.js
@@ -1,8 +1,0 @@
-import React from 'react'
-import './index.css'
-
-const Garage = () => {
-  return <div className="pagestyle_garage" />
-}
-
-export default Garage
