+++ src/components/Garage/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './index.css';
+
+const Garage = () => <div className="pagestyle_garage" />;
+
+export default Garage;
