--- src/components/Gas-stations/index.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import './index.css'
-
-const gas_stations = () => {
-  return (
-    <div className="pagestyle_gas">
-      <p>gas stations</p>
-    </div>
-  )
-}
-
-export default gas_stations
