+++ src/components/Gas-stations/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './index.css';
+
+const gasStations = () => (
+  <div className="pagestyle_gas">
+    <p>gas stations</p>
+  </div>
+);
+
+export default gasStations;
