--- src/components/Layout.js
@@ -1,4 +1,0 @@
-import React from 'react'
-import { Container } from 'react-bootstrap'
-
-export const Layout = (props) => <Container>{props.children}</Container>
