--- src/components/Paint-shops/index.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import './index.css'
-
-const PaintShops = () => {
-  return (
-    <div className="pagestyle_paint">
-      <p>paint shops</p>
-    </div>
-  )
-}
-
-export default PaintShops
