+++ src/components/Paint-shops/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './index.css';
+
+const PaintShops = () => (
+  <div className="pagestyle_paint">
+    <p>paint shops</p>
+  </div>
+);
+
+export default PaintShops;
